refactor(types): registerDimensions/Resources в MCE — снизить coupling MCS
